Add prop-types dependency, refactor index page structure, and create App and Layout components
