+++ package.json
@@ -10,7 +10,8 @@
     "contentful": "^10.1.6",
     "next": "13.4.12",
     "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "prop-types": "^15.8.1"
   },
   "devDependencies": {
     "@types/node": "24.0.3",
